refactor: move IssuedResourceService functionality
Move the functionality of `IssuedResourceService` to
`IssuanceService.completeCredentialIssuance`.

TIS21-4110
