+++ src/main/java/uk/nhs/hee/tis/trainee/credentials/api/IssueResource.java
@@ -41,7 +41,6 @@ import uk.nhs.hee.tis.trainee.credentials.dto.ProgrammeMembershipCredentialDto;
 import uk.nhs.hee.tis.trainee.credentials.dto.ProgrammeMembershipDataDto;
 import uk.nhs.hee.tis.trainee.credentials.mapper.CredentialDataMapper;
 import uk.nhs.hee.tis.trainee.credentials.service.IssuanceService;
-import uk.nhs.hee.tis.trainee.credentials.service.IssuedResourceService;
 
 /**
  * API endpoints for issuing trainee digital credentials.
@@ -53,13 +52,10 @@ public class IssueResource {
 
   private final IssuanceService service;
   private final CredentialDataMapper mapper;
-  private final IssuedResourceService issuedResourceService;
 
-  IssueResource(IssuanceService service, CredentialDataMapper mapper,
-      IssuedResourceService issuedResourceService) {
+  IssueResource(IssuanceService service, CredentialDataMapper mapper) {
     this.service = service;
     this.mapper = mapper;
-    this.issuedResourceService = issuedResourceService;
   }
 
   @PostMapping("/programme-membership")
@@ -118,7 +114,7 @@ public class IssueResource {
 
     log.info("Receiving callback for credential issuing.");
 
-    URI redirectUri = issuedResourceService.logIssuedResource(code, state, error, errorDescription);
+    URI redirectUri = service.completeCredentialVerification(code, state, error, errorDescription);
 
     log.info("Redirecting after credential issuing process.");
     return ResponseEntity.status(HttpStatus.FOUND).location(redirectUri).build();
