+++ src/main/java/uk/nhs/hee/tis/trainee/credentials/service/IssuanceService.java
@@ -23,21 +23,33 @@ package uk.nhs.hee.tis.trainee.credentials.service;
 
 import io.jsonwebtoken.Claims;
 import java.net.URI;
+import java.time.Instant;
 import java.util.Optional;
 import java.util.UUID;
+import lombok.extern.slf4j.Slf4j;
 import org.springframework.lang.Nullable;
 import org.springframework.stereotype.Service;
+import org.springframework.web.util.UriComponentsBuilder;
+import uk.nhs.hee.tis.trainee.credentials.config.GatewayProperties.IssuingProperties;
 import uk.nhs.hee.tis.trainee.credentials.dto.CredentialDto;
+import uk.nhs.hee.tis.trainee.credentials.dto.IssueResponseDto;
+import uk.nhs.hee.tis.trainee.credentials.mapper.CredentialMetadataMapper;
+import uk.nhs.hee.tis.trainee.credentials.model.CredentialMetadata;
+import uk.nhs.hee.tis.trainee.credentials.repository.CredentialMetadataRepository;
 
 /**
  * A service providing credential issuance functionality.
  */
+@Slf4j
 @Service
 public class IssuanceService {
 
   private final GatewayService gatewayService;
   private final JwtService jwtService;
   private final CachingDelegate cachingDelegate;
+  private final CredentialMetadataRepository credentialMetadataRepository;
+  private final CredentialMetadataMapper credentialMetadataMapper;
+  private final IssuingProperties properties;
 
   /**
    * Create a service providing credential issuance functionality.
@@ -47,10 +59,14 @@ public class IssuanceService {
    * @param cachingDelegate The caching delegate for caching data between requests.
    */
   IssuanceService(GatewayService gatewayService, JwtService jwtService,
-      CachingDelegate cachingDelegate) {
+      CachingDelegate cachingDelegate, CredentialMetadataRepository credentialMetadataRepository,
+      CredentialMetadataMapper credentialMetadataMapper, IssuingProperties properties) {
     this.gatewayService = gatewayService;
     this.jwtService = jwtService;
     this.cachingDelegate = cachingDelegate;
+    this.credentialMetadataRepository = credentialMetadataRepository;
+    this.credentialMetadataMapper = credentialMetadataMapper;
+    this.properties = properties;
   }
 
   /**
@@ -77,5 +93,68 @@ public class IssuanceService {
     cachingDelegate.cacheTraineeIdentifier(internalState, traineeId);
 
     return gatewayService.getCredentialUri(dto, nonce.toString(), internalState.toString());
+  }
+
+  /**
+   * Complete the credential issuance process.
+   *
+   * @param code             The code provided by the credential gateway.
+   * @param state            The state set in the initial gateway request.
+   * @param error            The error text, if the credential was not issued.
+   * @param errorDescription The error description, if the credential was not issued.
+   * @return The built redirect URI for completed issuance.
+   */
+  public URI completeCredentialVerification(String code, String state, String error,
+      String errorDescription) {
+    UUID stateUuid = UUID.fromString(state);
+    CredentialMetadata credentialMetadata = null;
+
+    if (error == null && code != null) {
+      log.info("Credential was issued successfully.");
+      URI tokenEndpoint = URI.create(properties.tokenEndpoint());
+      URI redirectEndpoint = URI.create(properties.redirectUri());
+      Claims claims = gatewayService.getTokenClaims(tokenEndpoint, redirectEndpoint, code, null);
+      UUID nonceUuid = UUID.fromString(claims.get("nonce", String.class));
+
+      Optional<String> traineeId = cachingDelegate.getTraineeIdentifier(stateUuid);
+      Optional<CredentialDto> credentialData = cachingDelegate.getCredentialData(nonceUuid);
+      if (traineeId.isPresent() && credentialData.isPresent()) {
+        IssueResponseDto issueResponseDto = fromIssuedResponse(claims, traineeId.get());
+        credentialMetadata = credentialMetadataMapper
+            .toCredentialMetadata(traineeId.get(), credentialData.get(), issueResponseDto);
+      }
+    } else {
+      log.info("Credential was not issued.");
+    }
+
+    if (credentialMetadata != null) {
+      credentialMetadataRepository.save(credentialMetadata);
+    }
+
+    Optional<String> clientState = cachingDelegate.getClientState(stateUuid);
+
+    // Build and return the redirect_uri
+    return UriComponentsBuilder.fromUriString("/credential-issued")
+        .queryParamIfPresent("state", clientState)
+        .queryParamIfPresent("error", Optional.ofNullable(error))
+        .queryParamIfPresent("error_description", Optional.ofNullable(errorDescription))
+        .build()
+        .toUri();
+  }
+
+  /**
+   * Build an issue response DTO from token claims and the user's authorization token.
+   *
+   * @param claims       The Claims data to map.
+   * @param traineeTisId The user's trainee TIS ID.
+   * @return The mapped credential log entry.
+   */
+
+  private IssueResponseDto fromIssuedResponse(Claims claims, String traineeTisId) {
+    String credentialId = claims.get("SerialNumber", String.class);
+    Instant issuedAt = Instant.ofEpochSecond(claims.get("iat", Long.class));
+    Instant expiresAt = Instant.ofEpochSecond(claims.get("exp", Long.class));
+
+    return new IssueResponseDto(credentialId, traineeTisId, issuedAt, expiresAt);
   }
 }
