+++ src/test/java/uk/nhs/hee/tis/trainee/credentials/api/IssueResourceTest.java
@@ -63,7 +63,6 @@ import uk.nhs.hee.tis.trainee.credentials.dto.ProgrammeMembershipCredentialDto;
 import uk.nhs.hee.tis.trainee.credentials.filter.FilterConfiguration;
 import uk.nhs.hee.tis.trainee.credentials.mapper.CredentialDataMapper;
 import uk.nhs.hee.tis.trainee.credentials.service.IssuanceService;
-import uk.nhs.hee.tis.trainee.credentials.service.IssuedResourceService;
 import uk.nhs.hee.tis.trainee.credentials.service.VerificationService;
 
 @WebMvcTest(IssueResource.class)
@@ -133,9 +132,6 @@ class IssueResourceTest {
 
   @MockBean
   private VerificationService verificationService;
-
-  @MockBean
-  private IssuedResourceService issuedResourceService;
 
   @MockBean
   private RestTemplateBuilder restTemplateBuilder;
@@ -471,8 +467,8 @@ class IssueResourceTest {
 
   @Test
   void shouldRedirectWhenIssuingCallbackCompleted() throws Exception {
-    when(issuedResourceService
-        .logIssuedResource(CODE_VALUE, STATE_VALUE, null, null))
+    when(issuanceService
+        .completeCredentialVerification(CODE_VALUE, STATE_VALUE, null, null))
         .thenReturn(URI.create("test-redirect"));
 
     mockMvc.perform(
