+++ src/test/java/uk/nhs/hee/tis/trainee/credentials/service/IssuanceServiceTest.java
@@ -22,39 +22,93 @@
 package uk.nhs.hee.tis.trainee.credentials.service;
 
 import static org.hamcrest.CoreMatchers.is;
+import static org.hamcrest.CoreMatchers.nullValue;
 import static org.hamcrest.MatcherAssert.assertThat;
+import static org.junit.jupiter.api.Assertions.assertEquals;
 import static org.mockito.ArgumentMatchers.any;
 import static org.mockito.ArgumentMatchers.eq;
 import static org.mockito.Mockito.mock;
 import static org.mockito.Mockito.verify;
+import static org.mockito.Mockito.verifyNoInteractions;
 import static org.mockito.Mockito.when;
 
+import io.jsonwebtoken.Claims;
 import io.jsonwebtoken.impl.DefaultClaims;
 import java.net.URI;
+import java.time.Instant;
+import java.time.LocalDate;
+import java.time.temporal.ChronoUnit;
+import java.util.Arrays;
 import java.util.Map;
 import java.util.Optional;
 import java.util.UUID;
+import java.util.stream.Collectors;
+import java.util.stream.Stream;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
+import org.junit.jupiter.params.ParameterizedTest;
+import org.junit.jupiter.params.provider.MethodSource;
+import org.mapstruct.factory.Mappers;
 import org.mockito.ArgumentCaptor;
 import uk.nhs.hee.tis.trainee.credentials.TestCredentialDto;
+import uk.nhs.hee.tis.trainee.credentials.config.GatewayProperties.IssuingProperties;
 import uk.nhs.hee.tis.trainee.credentials.dto.CredentialDto;
+import uk.nhs.hee.tis.trainee.credentials.dto.PlacementCredentialDto;
+import uk.nhs.hee.tis.trainee.credentials.dto.ProgrammeMembershipCredentialDto;
+import uk.nhs.hee.tis.trainee.credentials.mapper.CredentialMetadataMapper;
+import uk.nhs.hee.tis.trainee.credentials.model.CredentialMetadata;
+import uk.nhs.hee.tis.trainee.credentials.repository.CredentialMetadataRepository;
 
 class IssuanceServiceTest {
 
   private static final String AUTH_TOKEN = "dummy-auth-token";
 
+  private static final URI REDIRECT_URI = URI.create("https://credential.service/redirect-uri");
+  private static final URI TOKEN_ENDPOINT = URI.create("https://credential.gateway/token/endpoint");
+
+  private static final String CODE_VALUE = "some-code";
+  private static final String STATE_VALUE = UUID.randomUUID().toString();
+
+  private static final String TIS_ID = "the-tis-id";
+  private static final String TRAINEE_ID = "the-trainee-id";
+  private static final String CREDENTIAL_ID = "123-456-789";
+  private static final Instant ISSUED_AT = Instant.MIN.truncatedTo(ChronoUnit.SECONDS);
+  private static final Instant EXPIRES_AT = Instant.MAX.truncatedTo(ChronoUnit.SECONDS);
+
   private IssuanceService issuanceService;
   private GatewayService gatewayService;
   private JwtService jwtService;
   private CachingDelegate cachingDelegate;
+  private CredentialMetadataRepository credentialMetadataRepository;
 
   @BeforeEach
   void setUp() {
     gatewayService = mock(GatewayService.class);
     jwtService = mock(JwtService.class);
     cachingDelegate = mock(CachingDelegate.class);
-    issuanceService = new IssuanceService(gatewayService, jwtService, cachingDelegate);
+    credentialMetadataRepository = mock(CredentialMetadataRepository.class);
+    var credentialMetadataMapper = Mappers.getMapper(CredentialMetadataMapper.class);
+
+    var properties = new IssuingProperties("", "", TOKEN_ENDPOINT.toString(), null,
+        REDIRECT_URI.toString());
+
+    issuanceService = new IssuanceService(gatewayService, jwtService, cachingDelegate,
+        credentialMetadataRepository, credentialMetadataMapper, properties);
+  }
+
+  /**
+   * Create a stream of credentials, allowing testing of common functionality with different
+   * credential types.
+   *
+   * @return The stream of credential objects.
+   */
+  private static Stream<CredentialDto> credentialMethodSource() {
+    return Stream.of(
+        new TestCredentialDto(TIS_ID),
+        new PlacementCredentialDto(TIS_ID, "", "", "", "", "", LocalDate.MIN, LocalDate.MAX),
+        new ProgrammeMembershipCredentialDto(TIS_ID, "", LocalDate.MIN, LocalDate.MAX)
+
+    );
   }
 
   @Test
@@ -136,5 +190,99 @@ class IssuanceServiceTest {
     Optional<URI> uri = issuanceService.startCredentialIssuance(AUTH_TOKEN, credentialData, null);
 
     assertThat("Unexpected URI.", uri, is(Optional.of(redirectUri)));
+  }
+
+  @Test
+  void shouldNotSaveToRepositoryWhenCredentialNotIssued() {
+    issuanceService.completeCredentialVerification(null, STATE_VALUE, "error", "error_description");
+    verifyNoInteractions(credentialMetadataRepository);
+  }
+
+  @ParameterizedTest
+  @MethodSource("credentialMethodSource")
+  void shouldSaveToRepositoryWhenCredentialIssued(CredentialDto credentialData) {
+    UUID nonce = UUID.randomUUID();
+    Claims claimsIssued = new DefaultClaims();
+    claimsIssued.put("nonce", nonce.toString());
+    claimsIssued.put("SerialNumber", CREDENTIAL_ID);
+    claimsIssued.put("iat", ISSUED_AT.getEpochSecond());
+    claimsIssued.put("exp", EXPIRES_AT.getEpochSecond());
+
+    // TODO: this is quite brittle
+    when(gatewayService.getTokenClaims(eq(TOKEN_ENDPOINT), eq(REDIRECT_URI), eq(CODE_VALUE), any()))
+        .thenReturn(claimsIssued);
+    when(cachingDelegate.getCredentialData(nonce)).thenReturn(Optional.of(credentialData));
+    when(cachingDelegate.getTraineeIdentifier(UUID.fromString(STATE_VALUE))).thenReturn(
+        Optional.of(TRAINEE_ID));
+    issuanceService.completeCredentialVerification(CODE_VALUE, STATE_VALUE, null, null);
+
+    ArgumentCaptor<CredentialMetadata> argument = ArgumentCaptor.forClass(CredentialMetadata.class);
+    verify(credentialMetadataRepository).save(argument.capture());
+    assertEquals(CREDENTIAL_ID, argument.getValue().getCredentialId());
+    assertEquals(credentialData.getScope(), argument.getValue().getCredentialType());
+    assertEquals(TIS_ID, argument.getValue().getTisId());
+    assertEquals(TRAINEE_ID, argument.getValue().getTraineeId());
+    assertEquals(ISSUED_AT, argument.getValue().getIssuedAt());
+    assertEquals(EXPIRES_AT, argument.getValue().getExpiresAt());
+  }
+
+  @Test
+  void shouldReturnCredentialIssuedWhenGatewayIssueSuccessful() {
+    URI uri = issuanceService.completeCredentialVerification(null, STATE_VALUE, null, null);
+
+    assertThat("Unexpected URI path.", uri.getPath(), is("/credential-issued"));
+  }
+
+  @Test
+  void shouldIncludeGatewayErrorsInFinalRedirectWhenHasError() {
+    URI uri = issuanceService.completeCredentialVerification(null, STATE_VALUE, "error_code",
+        "error description");
+
+    Map<String, String> queryParams = splitQueryParams(uri);
+    assertThat("Unexpected query parameter count.", queryParams.size(), is(2));
+    assertThat("Unexpected error.", queryParams.get("error"), is("error_code"));
+    assertThat("Unexpected error description.", queryParams.get("error_description"),
+        is("error description"));
+  }
+
+  @Test
+  void shouldNotIncludeGatewayErrorsInFinalRedirectWhenNoError() {
+    URI uri = issuanceService.completeCredentialVerification(null, STATE_VALUE, null, null);
+
+    assertThat("Unexpected uri query.", uri.getQuery(), nullValue());
+  }
+
+  @Test
+  void shouldIncludeClientStateInFinalRedirectWhenGiven() {
+    when(cachingDelegate.getClientState(UUID.fromString(STATE_VALUE))).thenReturn(
+        Optional.of("some-client-state"));
+
+    URI uri = issuanceService.completeCredentialVerification(null, STATE_VALUE, null, null);
+
+    Map<String, String> queryParams = splitQueryParams(uri);
+    assertThat("Unexpected query parameter count.", queryParams.size(), is(1));
+    assertThat("Unexpected error.", queryParams.get("state"), is("some-client-state"));
+  }
+
+  @Test
+  void shouldNotIncludeClientStateInFinalRedirectWhenNotGiven() {
+    when(cachingDelegate.getClientState(UUID.fromString(STATE_VALUE))).thenReturn(Optional.empty());
+
+    URI uri = issuanceService.completeCredentialVerification(null, STATE_VALUE, null, null);
+
+    assertThat("Unexpected uri query.", uri.getQuery(), nullValue());
+  }
+
+  /**
+   * Split the query params of a URI in to a Map.
+   *
+   * @param uri The URI to split the query params from.
+   * @return A map of parameter names to value.
+   */
+  private Map<String, String> splitQueryParams(URI uri) {
+    return Arrays.stream(
+            uri.getQuery().split("&"))
+        .map(param -> param.split("="))
+        .collect(Collectors.toMap(keyValue -> keyValue[0], keyValue -> keyValue[1]));
   }
 }
